requirements.txt
azure_storage
streamlit